+++ src/command_modules/azure-cli-vm/azure/cli/command_modules/vm/_validators.py
@@ -6,6 +6,10 @@
 # pylint:disable=too-many-lines
 
 import os
+try:
+    from urllib.parse import urlparse
+except ImportError:
+    from urlparse import urlparse  # pylint: disable=import-error
 
 from knack.log import get_logger
 from knack.util import CLIError
@@ -13,7 +17,7 @@ from knack.util import CLIError
 from azure.cli.core.commands.validators import (
     get_default_location_from_resource_group, validate_file_or_dict, validate_parameter_set, validate_tags)
 from azure.cli.core.util import hash_string
-from azure.cli.command_modules.vm._vm_utils import check_existence, get_target_network_api
+from azure.cli.command_modules.vm._vm_utils import check_existence, get_target_network_api, get_storage_blob_uri
 from azure.cli.command_modules.vm._template_builder import StorageProfile
 import azure.cli.core.keys as keys
 
@@ -198,15 +202,11 @@ def _parse_image_argument(cmd, namespace):
     from msrestazure.azure_exceptions import CloudError
     import re
 
-    # 1 - easy check for URI
-    if namespace.image.lower().endswith('.vhd'):
-        return 'uri'
-
-    # 2 - check if a fully-qualified ID (assumes it is an image ID)
+    # 1 - check if a fully-qualified ID (assumes it is an image ID)
     if is_valid_resource_id(namespace.image):
         return 'image_id'
 
-    # 3 - attempt to match an URN pattern
+    # 2 - attempt to match an URN pattern
     urn_match = re.match('([^:]*):([^:]*):([^:]*):([^:]*)', namespace.image)
     if urn_match:
         namespace.os_publisher = urn_match.group(1)
@@ -222,6 +222,10 @@ def _parse_image_argument(cmd, namespace):
                 namespace.plan_publisher = image_plan.publisher
 
         return 'urn'
+
+    # 3 - unmanaged vhd based images?
+    if urlparse(namespace.image).scheme:
+        return 'uri'
 
     # 4 - attempt to match an URN alias (most likely)
     from azure.cli.command_modules.vm._actions import load_images_from_aliases_doc
@@ -424,15 +428,13 @@ def _validate_vm_create_storage_profile(cmd, namespace, for_scale_set=False):
         namespace.os_type = 'windows' if 'windows' in namespace.os_offer.lower() else 'linux'
 
     from ._vm_utils import normalize_disk_info
-    # accelnet and attach_data_disks are not exposed yet for VMSS, so use 'getattr' to avoid crash
-    write_accelerator_settings = getattr(namespace, 'write_accelerator', None)
+    # attach_data_disks are not exposed yet for VMSS, so use 'getattr' to avoid crash
     namespace.disk_info = normalize_disk_info(image_data_disks=image_data_disks,
                                               data_disk_sizes_gb=namespace.data_disk_sizes_gb,
                                               attach_data_disks=getattr(namespace, 'attach_data_disks', []),
                                               storage_sku=namespace.storage_sku,
                                               os_disk_caching=namespace.os_caching,
-                                              data_disk_cachings=namespace.data_caching,
-                                              write_accelerator_settings=write_accelerator_settings)
+                                              data_disk_cachings=namespace.data_caching)
 
 
 def _validate_vm_create_storage_account(cmd, namespace):
@@ -893,7 +895,8 @@ def process_vm_create_namespace(cmd, namespace):
     if namespace.license_type and namespace.os_type.lower() != 'windows':
         raise CLIError('usage error: --license-type is only applicable on Windows VM')
     _validate_vm_vmss_msi(cmd, namespace)
-
+    if namespace.boot_diagnostics_storage:
+        namespace.boot_diagnostics_storage = get_storage_blob_uri(cmd.cli_ctx, namespace.boot_diagnostics_storage)
 # endregion
 
 
@@ -1148,7 +1151,7 @@ def _figure_out_storage_source(cli_ctx, resource_group_name, source):
     source_blob_uri = None
     source_disk = None
     source_snapshot = None
-    if source.lower().endswith('.vhd'):
+    if urlparse(source).scheme:  # a uri?
         source_blob_uri = source
     elif '/disks/' in source.lower():
         source_disk = source