+++ src/command_modules/azure-cli-acs/azure/cli/command_modules/acs/custom.py
@@ -1430,6 +1430,87 @@ def aks_upgrade(cmd, client, resource_group_name, name, kubernetes_version, no_w
     return sdk_no_wait(no_wait, client.create_or_update, resource_group_name, name, instance)
 
 
+DEV_SPACES_EXTENSION_NAME = 'dev-spaces-preview'
+DEV_SPACES_EXTENSION_MODULE = 'azext_dev_spaces_preview.custom'
+
+
+def aks_use_dev_spaces(cmd, client, name, resource_group_name, space_name='default', parent_space_name=None):
+    """
+    Use Azure Dev Spaces with a managed Kubernetes cluster.
+
+    :param name: Name of the managed cluster.
+    :type name: String
+    :param resource_group_name: Name of resource group. You can configure the default group. \
+    Using 'az configure --defaults group=<name>'.
+    :type resource_group_name: String
+    :param space_name: Name of the dev space to use.
+    :type space_name: String
+    :param parent_space_name: Name of a parent dev space to inherit from when creating a new dev space. \
+    By default, if there is already a single dev space with no parent, the new space inherits from this one.
+    :type parent_space_name: String
+    """
+
+    if _get_or_add_extension(DEV_SPACES_EXTENSION_NAME):
+        azext_custom = _get_azext_module(DEV_SPACES_EXTENSION_NAME, DEV_SPACES_EXTENSION_MODULE)
+        try:
+            azext_custom.ads_use_dev_spaces(name, resource_group_name, space_name, parent_space_name)
+        except AttributeError as ae:
+            raise CLIError(ae)
+
+
+def aks_remove_dev_spaces(cmd, client, name, resource_group_name, prompt=False):
+    """
+    Remove Azure Dev Spaces from a managed Kubernetes cluster.
+
+    :param name: Name of the managed cluster.
+    :type name: String
+    :param resource_group_name: Name of resource group. You can configure the default group. \
+    Using 'az configure --defaults group=<name>'.
+    :type resource_group_name: String
+    :param prompt: Do not prompt for confirmation.
+    :type prompt: bool
+    """
+
+    if _get_or_add_extension(DEV_SPACES_EXTENSION_NAME):
+        azext_custom = _get_azext_module(DEV_SPACES_EXTENSION_NAME, DEV_SPACES_EXTENSION_MODULE)
+        try:
+            azext_custom.ads_remove_dev_spaces(name, resource_group_name, prompt)
+        except AttributeError as ae:
+            raise CLIError(ae)
+
+
+def _get_azext_module(extension_name, module_name):
+    try:
+        # Adding the installed extension in the path
+        from azure.cli.core.extension import get_extension_path
+        ext_dir = get_extension_path(extension_name)
+        sys.path.append(ext_dir)
+        # Import the extension module
+        from importlib import import_module
+        azext_custom = import_module(module_name)
+        return azext_custom
+    except ImportError as ie:
+        raise CLIError(ie)
+
+
+def _install_dev_spaces_extension(extension_name):
+    try:
+        from azure.cli.command_modules.extension import custom
+        custom.add_extension(extension_name=extension_name)
+    except Exception:  # nopa pylint: disable=broad-except
+        return False
+    return True
+
+
+def _get_or_add_extension(extension_name):
+    from azure.cli.core.extension import (ExtensionNotInstalledException, get_extension)
+    try:
+        get_extension(extension_name)
+    except ExtensionNotInstalledException:
+        return _install_dev_spaces_extension(extension_name)
+    return True
+
+
 def _ensure_aks_service_principal(cli_ctx,
                                   service_principal=None,
                                   client_secret=None,