+++ src/command_modules/azure-cli-acs/HISTORY.rst
@@ -6,6 +6,7 @@ Release History
 2.0.33
 ++++++
 * add new Dev-Spaces commands: `az aks use-dev-spaces` and `az aks remove-dev-spaces`
+* fix typo in help message
 
 2.0.32
 ++++++